Validator earnings are now split into 3 categories: block production, transaction attestation and mpc participation where each category can earn fees or get monetary/gas penalties

+++ src/tangent/validator/storage/chainstate.h
@@ -130,7 +130,7 @@ namespace tangent
 
 		struct factor_index_window final : factor_window
 		{
-			vector<size_t> indices;
+			vector<uint64_t> indices;
 
 			uint8_t type() const override
 			{
